post-command: Fix keyless signing. (#7)
The keyless signing feature (used to) rely on the COSIGN_EXPERIMENTAL
environment variable being set to "1". The post-command shell script
attempts to set this environment variable when executing cosign if
the user specifies keyless signing.

The environment variable name was missing the underscore character.
This caused keyless signing attempts to fail because the shell would
interpret "COSIGN" as a shell command. [1]

This commit also removes some extra empty space.

References

1. https://buildkite.com/metal-toolbox/governor-slack-addon/builds/106#018c122e-21ab-4488-b03a-0c5131595448